chore: move Icon to elements
