+++ src/admin/components/elements/BlockEditor/menus/ContentItemMenu/index.tsx
@@ -5,7 +5,7 @@ import { Editor } from '@tiptap/react';
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DropdownButton } from '../../ui/DropdownButton/index.js';
-import { Icon } from '../../ui/Icon/index.js';
+import { Icon } from '../../../Icon/index.js';
 import { ToolbarButton } from '../../ui/ToolbarButton/index.js';
 import { useContentItemActions } from './hooks/useContentItemActions.js';
 import { useData } from './hooks/useData.js';
